requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
funcsigs==1.0.2
gast==0.3.2
grpcio==1.24.0
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.17.2
protobuf==3.9.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
texar==0.2.1
urllib3==1.25.6
Werkzeug==0.16.0
